# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/swap/quote": {
      "get": {
        "operationId": "SwapController_getQuote",
        "parameters": [
          {
            "name": "fromToken",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "USDC",
                "CNGN"
              ],
              "type": "string"
            }
          },
          {
            "name": "toToken",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "USDC",
                "CNGN"
              ],
              "type": "string"
            }
          },
          {
            "name": "amount",
            "required": true,
            "in": "query",
            "description": "Amount of fromToken to swap",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quote with source/destination amounts and exchange rate"
          }
        },
        "summary": "Get a live swap quote between USDC and CNGN",
        "tags": [
          "Swap"
        ],
        "security": [
          {
            "API-Key": []
          }
        ]
      }
    },
    "/swap/balance/{token}/{address}": {
      "get": {
        "operationId": "SwapController_getTokenBalance",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Token symbol (e.g., USDC, CNGN)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": true,
            "in": "path",
            "description": "Stellar public key (G...)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token balance as string"
          }
        },
        "summary": "Get token balance for an address",
        "tags": [
          "Swap"
        ],
        "security": [
          {
            "API-Key": []
          }
        ]
      }
    },
    "/swap/trustline/{token}/{address}": {
      "get": {
        "operationId": "SwapController_hasTrustline",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Token symbol (e.g., USDC, CNGN)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": true,
            "in": "path",
            "description": "Stellar public key (G...)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Whether the trustline exists"
          }
        },
        "summary": "Check whether an address has a trustline for a token",
        "tags": [
          "Swap"
        ],
        "security": [
          {
            "API-Key": []
          }
        ]
      }
    },
    "/swap/trustline/{token}/{address}/sponsor": {
      "post": {
        "description": "Builds a transaction (partially signed by AutoRamp's distribution account) that creates a trustline for the given token, with AutoRamp sponsoring the reserve. The user's wallet must sign and submit it. Requires authentication.",
        "operationId": "SwapController_getSponsoredTrustline",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "Token symbol (e.g., USDC, CNGN)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "required": true,
            "in": "path",
            "description": "Stellar public key (G...) that will hold the trustline; must already exist on-chain",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Unsigned (by the user) transaction XDR + network passphrase"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "API-Key": []
          }
        ],
        "summary": "Get a sponsored trustline transaction to sign",
        "tags": [
          "Swap"
        ]
      }
    },
    "/swap/balances": {
      "get": {
        "operationId": "SwapController_getTokenBalances",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "query",
            "description": "Stellar public key (G...)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Record of balances"
          }
        },
        "summary": "Get balances for USDC and CNGN for an address",
        "tags": [
          "Swap"
        ],
        "security": [
          {
            "API-Key": []
          }
        ]
      }
    },
    "/swap/usd-ngn-rate": {
      "get": {
        "operationId": "SwapController_getUsdNgnRate",
        "parameters": [],
        "responses": {
          "200": {
            "description": "USD/NGN rate"
          }
        },
        "summary": "Get USD/NGN rate from MonieRate API",
        "tags": [
          "Swap"
        ],
        "security": [
          {
            "API-Key": []
          }
        ]
      }
    },
    "/swap/estimate-ngn": {
      "get": {
        "operationId": "SwapController_estimateNgn",
        "parameters": [
          {
            "name": "cngnAmount",
            "required": true,
            "in": "query",
            "description": "Amount in CNGN",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estimated NGN value with rates"
          }
        },
        "summary": "Calculate estimated NGN value from CNGN amount",
        "tags": [
          "Swap"
        ],
        "security": [
          {
            "API-Key": []
          }
        ]
      }
    },
    "/swap/initialize": {
      "post": {
        "description": "Initializes offramp first to get recipient wallet address, then creates swap record. Returns recipient address and swap parameters for frontend to execute the swap. Requires authentication.",
        "operationId": "SwapController_initializeSwap",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitializeSwapDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Swap and offramp initialization response with recipient address"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing authentication"
          }
        },
        "security": [
          {
            "API-Key": []
          }
        ],
        "summary": "Initialize swap transaction (USDC → CNGN → Offramp)",
        "tags": [
          "Swap"
        ]
      }
    },
    "/swap/{reference}/complete": {
      "post": {
        "description": "Updates swap transaction status after user executes the swap on blockchain. Requires authentication.",
        "operationId": "SwapController_updateSwapAfterExecution",
        "parameters": [
          {
            "name": "reference",
            "required": true,
            "in": "path",
            "description": "Swap transaction reference",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transactionHash": {
                    "type": "string",
                    "example": "0x..."
                  },
                  "sourceAddress": {
                    "type": "string",
                    "example": "0x..."
                  }
                },
                "required": [
                  "transactionHash",
                  "sourceAddress"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Swap transaction updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "API-Key": []
          }
        ],
        "summary": "Update swap transaction after execution",
        "tags": [
          "Swap"
        ]
      }
    },
    "/swap/create": {
      "post": {
        "description": "Creates a swap transaction record in the database. The actual swap execution happens on-chain via the frontend. Requires authentication.",
        "operationId": "SwapController_createSimpleSwap",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSimpleSwapDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Swap transaction created successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "API-Key": []
          }
        ],
        "summary": "Create a simple swap transaction (store in database)",
        "tags": [
          "Swap"
        ]
      }
    },
...
  },
  "info": {
    "title": "CNGN Ramp Business API",
    "description": "Programmatic API for businesses to perform onramp, offramp, and swap operations.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Your business API key generated from the admin portal."
      }
    },
    "schemas": {
...
      "OfframpDestinationDto": {
        "type": "object",
        "properties": {
          "bankCode": {
            "type": "string",
            "example": "058",
            "description": "Bank code"
          },
          "accountNumber": {
            "type": "string",
            "example": "1234567890",
            "description": "Account number"
          }
        },
        "required": [
          "bankCode",
          "accountNumber"
        ]
      },
      "InitializeSwapDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 140196,
            "description": "NGN amount for offramp (estimated NGN value)"
          },
          "usdcAmount": {
            "type": "number",
            "example": 100,
            "description": "USDC amount for swap (amount user wants to swap)"
          },
          "slippage": {
            "type": "number",
            "example": 0.05,
            "description": "Slippage tolerance (0-1, e.g., 0.05 = 5%)"
          },
          "offrampDestination": {
            "description": "Offramp destination (bank details)",
            "allOf": [
              {
                "$ref": "#/components/schemas/OfframpDestinationDto"
              }
            ]
          },
          "network": {
            "type": "string",
            "example": "stellar",
            "description": "Network (default: stellar)"
          }
        },
        "required": [
          "amount",
          "usdcAmount",
          "slippage",
          "offrampDestination"
        ]
      },
      "CreateSimpleSwapDto": {
        "type": "object",
        "properties": {
          "fromTokenType": {
            "type": "string",
            "example": "USDC",
            "description": "Token type to swap from (USDC or CNGN)",
            "enum": [
              "USDC",
              "CNGN"
            ]
          },
          "toTokenType": {
            "type": "string",
            "example": "CNGN",
            "description": "Token type to swap to (USDC or CNGN)",
            "enum": [
              "USDC",
              "CNGN"
            ]
          },
          "fromAmount": {
            "type": "number",
            "example": 100.5,
            "description": "Amount to swap from"
          },
          "toAmount": {
            "type": "number",
            "example": 100.5,
            "description": "Estimated amount to receive"
          },
          "exchangeRate": {
            "type": "number",
            "example": 1,
            "description": "Exchange rate (fromAmount/toAmount)"
          },
          "sourceAddress": {
            "type": "string",
            "example": "GDQP2KPQGKIHYJGXNUIYOMHARUARCA7DJT5FO2FFOOKY3B2WSQHG4W37",
            "description": "Source wallet address (user's Stellar public key)"
          },
          "destinationAddress": {
            "type": "string",
            "example": "GDQP2KPQGKIHYJGXNUIYOMHARUARCA7DJT5FO2FFOOKY3B2WSQHG4W37",
            "description": "Destination wallet address (user's wallet, same as source for simple swaps)"
          },
          "network": {
            "type": "string",
            "example": "stellar",
            "description": "Network (defaults to stellar)"
          },
          "slippage": {
            "type": "number",
            "example": 0.05,
            "description": "Slippage tolerance (defaults to 0.05 = 5%)"
          }
        },
        "required": [
          "fromTokenType",
          "toTokenType",
          "fromAmount",
          "toAmount",
          "exchangeRate",
          "sourceAddress",
          "destinationAddress"
        ]
      },
...
    }
  }
}